[Refactor] StyleSheet 정리
