+++ src/layouts/RootLayout.js
@@ -1,14 +1,14 @@
 import { Outlet } from 'react-router-dom';
 
-import Header from '@components/root/MainHeader';
-import Aside from '@components/root/MainAside';
+import MainHeader from '@components/root/MainHeader';
+import MainAside from '@components/root/MainAside';
 
 const RootLayout = () => {
   return (
     <>
-      <Header />
+      <MainHeader />
 
-      <Aside />
+      <MainAside />
 
       <main style={{ padding: '8px 0 128px' }}>
         <Outlet />
