+++ src/pages/Detail.js
@@ -2,7 +2,6 @@ import { useSelector, shallowEqual } from 'react-redux';
 import { useParams } from 'react-router-dom';
 
 import PostDetail from '@components/detail/PostDetail';
-import StyledIndex from '@styles/main/Index-styled';
 
 const Detail = () => {
   const { postId } = useParams();
@@ -13,11 +12,7 @@ const Detail = () => {
     return null;
   }
 
-  return (
-    <StyledIndex>
-      <PostDetail post={renderPost[0]} />
-    </StyledIndex>
-  );
+  return <PostDetail post={renderPost[0]} />;
 };
 
 export default Detail;
