+++ src/styles/components/common/PostForm-styled.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-const StyledCreate = styled.section`
+const StyledPostForm = styled.section`
   width: 800px;
   margin: 32px auto 0;
 
@@ -114,60 +114,6 @@ const StyledCreate = styled.section`
     .footer {
       position: relative;
 
-      .tag-list {
-        position: absolute;
-        top: 0;
-        left: 84px;
-
-        display: flex;
-        flex-direction: row;
-        justify-content: start;
-
-        .tag-item {
-          border: 1px solid #ccc;
-
-          display: flex;
-          align-items: center;
-          justify-content: space-between;
-          margin: 3px;
-          padding: 3px;
-          background-color: tomato;
-          border-radius: 5px;
-          color: white;
-
-          font-size: 10px;
-
-          cursor: pointer;
-
-          &:hover {
-            background-color: gray;
-          }
-        }
-      }
-
-      .tag-box {
-        display: flex;
-        align-items: center;
-        gap: 16px;
-
-        height: 64px;
-        padding: 16px 24px 0;
-        margin: 0 16px;
-
-        background-color: #f8f8f8;
-
-        input {
-          flex-grow: 1;
-          padding-inline: 12px;
-          height: 36px;
-          border: 1px solid #d9d9d9;
-
-          &:focus {
-            outline: none;
-          }
-        }
-      }
-
       .post-button-container {
         display: flex;
         justify-content: space-between;
@@ -210,4 +156,4 @@ const StyledCreate = styled.section`
   }
 `;
 
-export default StyledCreate;
+export default StyledPostForm;
