+++ src/styles/components/detail/CategoryList-styled.js
