+++ src/styles/components/detail/PostContent-styled.js
