+++ src/styles/components/home/Search-styled.js
