+++ src/styles/layouts/AuthLayout-styled.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+const StyledAuthLayout = styled.section`
+  padding-top: 128px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  .title {
+    margin-bottom: 48px;
+    text-transform: uppercase;
+  }
+`;
+
+export default StyledAuthLayout;
