--- src/styles/main/Index-styled.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const StyledIndex = styled.div`
-  padding: 8px 0 128px;
-`;
-
-export default StyledIndex;
