+++ src/styles/pages/Error-styled.js
